Fix signal icon is not shown
The flag GSM of signal strength is set as false under OOS, if lower
layer does not notify signal strength again after in-service, the
signal level is based on CDMA, so that signal icon is unable to be
shown.

To fix it, change the preset flag GSM based on phone type.

Change-Id: Ib565bcf7afbf1e8b2771a35b0e9c87ab7ffd6301
Signed-off-by: Josh Fox (XlxFoXxlX) <joshfox87@gmail.com>